Run the checks this host can actually run
`verify` reported clang-tidy "not on PATH — brew install llvm" on a host that
has llvm installed: Homebrew keeps it keg-only, so `which` answers nothing and
the remedy was an instruction to redo what was already done. It now looks in
the keg-only prefix too, and puts it on PATH for the run. The SDL3 headers are
resolved from the host package when BENEFACTOR_SDL3_DIR names no checkout,
rather than leaving tidy to report `SDL3/SDL.h` not found as if the sources
were broken.

Running it then found real findings. `bugprone-signed-bitwise` counts a plain
`8` shift distance as a signed operand from clang-tidy 20 on, which makes every
hardware-word shift a finding; the check keeps reporting signed values, with
positive literals ignored. The promotions it flagged in port.h and
guest_runtime.cpp are widened explicitly, and a test's main says why an
exception leaving it is the failure being reported.

+++ src/port/port.h
@@ -10,21 +10,23 @@ extern uint8_t *g_chip;
 
 /* Read/write helpers for chip RAM */
 static inline uint8_t r8(uint32_t a) { return g_chip[a]; }
-static inline uint16_t r16(uint32_t a) { return (g_chip[a] << 8) | g_chip[a + 1]; }
+static inline uint16_t r16(uint32_t a) {
+    return (uint16_t)(((uint32_t)g_chip[a] << 8) | (uint32_t)g_chip[a + 1]);
+}
 static inline uint32_t r32(uint32_t a) {
     return ((uint32_t)g_chip[a] << 24) | ((uint32_t)g_chip[a + 1] << 16) |
-           ((uint32_t)g_chip[a + 2] << 8) | g_chip[a + 3];
+           ((uint32_t)g_chip[a + 2] << 8) | (uint32_t)g_chip[a + 3];
 }
 static inline void w8(uint32_t a, uint8_t v) { g_chip[a] = v; }
 static inline void w16(uint32_t a, uint16_t v) {
-    g_chip[a] = v >> 8;
-    g_chip[a + 1] = v;
+    g_chip[a] = (uint8_t)((uint32_t)v >> 8);
+    g_chip[a + 1] = (uint8_t)v;
 }
 static inline void w32(uint32_t a, uint32_t v) {
-    g_chip[a] = v >> 24;
-    g_chip[a + 1] = v >> 16;
-    g_chip[a + 2] = v >> 8;
-    g_chip[a + 3] = v;
+    g_chip[a] = (uint8_t)(v >> 24);
+    g_chip[a + 1] = (uint8_t)(v >> 16);
+    g_chip[a + 2] = (uint8_t)(v >> 8);
+    g_chip[a + 3] = (uint8_t)v;
 }
 
 /* Screen dimensions */
